pyproject.toml
[tool.poetry]
name = "selfie_stabilizer"
version = "0.1.0"
description = "Converts a folder of selfie images into a stabilized video."
authors = ["Jaime <jaimeliew1@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
dlib-bin = "^19.22.0"
scipy = "^1.8.0"
numpy = "^1.22.2"
tqdm = "^4.62.3"
imutils = "^0.5.4"
opencv-python = "^4.5.5"
matplotlib = "^3.5.1"


[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
icecream = "^2.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
